add a custom hook for the csv parsing files to json
