+++ src/Hooks/getData.js/index.js
@@ -1,8 +1,18 @@
-import React from 'react'
+import Papa from "papaparse";
+import { useEffect, useState } from 'react'
 
-export const UseGetData = (url, ) => {
-  return (
-    <>
-    </>
-  )
+export const UseGetData = (csv) => {
+  const [file, setFile] = useState([]);
+
+  useEffect(() => {
+    const data = csv;
+    Papa.parse(data, {
+      download: true,
+      header: true,
+      complete: function (result) {
+        setFile(result.data);
+      },
+    });
+  }, [csv]);
+  return file
 }
